Refactors autolinks configuration  - Updates autolink interfaces for better type safety  - Enhances autolink configuration with explicit boolean flags  - Adjusts remote providers to use updated autolink structure
