+++ src/git/remotes/azure-devops.ts
@@ -1,6 +1,5 @@
 import type { Range, Uri } from 'vscode';
-import type { DynamicAutolinkReference } from '../../annotations/autolinks';
-import type { AutolinkReference } from '../../config';
+import type { AutolinkReference, DynamicAutolinkReference } from '../../autolinks';
 import type { GkProviderId } from '../../gk/models/repositoryIdentities';
 import type { Brand, Unbrand } from '../../system/brand';
 import type { Repository } from '../models/repository';
@@ -65,6 +64,8 @@ export class AzureDevOpsRemote extends RemoteProvider {
 				{
 					prefix: '#',
 					url: `${workUrl}/_workitems/edit/<num>`,
+					alphanumeric: false,
+					ignoreCase: false,
 					title: `Open Work Item #<num> on ${this.name}`,
 
 					type: 'issue',
@@ -74,6 +75,8 @@ export class AzureDevOpsRemote extends RemoteProvider {
 					// Default Pull request message when merging a PR in ADO. Will not catch commits & pushes following a different pattern.
 					prefix: 'PR ',
 					url: `${this.baseUrl}/pullrequest/<num>`,
+					alphanumeric: false,
+					ignoreCase: false,
 					title: `Open Pull Request #<num> on ${this.name}`,
 
 					type: 'pullrequest',
