+++ src/webviews/commitDetails/commitDetailsWebview.ts
@@ -2,8 +2,8 @@ import { EntityIdentifierUtils } from '@gitkraken/provider-apis';
 import type { CancellationToken, ConfigurationChangeEvent, TextDocumentShowOptions } from 'vscode';
 import { CancellationTokenSource, Disposable, env, Uri, window } from 'vscode';
 import { extractDraftMessage } from '../../ai/aiProviderService';
-import type { MaybeEnrichedAutolink } from '../../annotations/autolinks';
-import { serializeAutolink } from '../../annotations/autolinks';
+import type { MaybeEnrichedAutolink } from '../../autolinks';
+import { serializeAutolink } from '../../autolinks';
 import { getAvatarUri } from '../../avatars';
 import type { CopyMessageToClipboardCommandArgs } from '../../commands/copyMessageToClipboard';
 import type { CopyShaToClipboardCommandArgs } from '../../commands/copyShaToClipboard';
